+++ apps/minifront/src/components/ibc/ibc-out/ibc-out-form.tsx
@@ -64,6 +64,11 @@ export const IbcOutForm = () => {
             issue: 'insufficient funds',
             checkFn: () => validationErrors.amountErr,
           },
+          {
+            type: 'error',
+            issue: 'invalid decimal length',
+            checkFn: () => validationErrors.exponentErr,
+          },
         ]}
         balances={filteredBalances}
       />