+++ apps/minifront/src/components/shared/gas-fee.tsx
@@ -1,14 +1,14 @@
 import {
   Fee,
   FeeTier_Tier,
-} from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/component/fee/v1/fee_pb.js';
+} from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/component/fee/v1/fee_pb';
 import { SegmentedPicker, SegmentedPickerOption } from '@repo/ui/components/ui/segmented-picker';
 import { InputBlock } from './input-block';
 import { ValueViewComponent } from '@repo/ui/components/ui/value';
 import {
   Metadata,
   ValueView,
-} from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb.js';
+} from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
 
 const FEE_TIER_OPTIONS: SegmentedPickerOption<FeeTier_Tier>[] = [
   {
@@ -30,23 +30,23 @@ export const GasFee = ({
   feeTier,
   stakingAssetMetadata,
   setFeeTier,
+  assetFeeMetadata,
 }: {
   fee: Fee | undefined;
   feeTier: FeeTier_Tier;
   stakingAssetMetadata?: Metadata;
+  assetFeeMetadata?: Metadata;
   setFeeTier: (feeTier: FeeTier_Tier) => void;
 }) => {
-  if (!stakingAssetMetadata) {
-    return null;
-  }
+  // If the metadata for the fee asset is undefined, fallback to using the bundled staking asset metadata.
+  const feeMetadata = assetFeeMetadata ?? stakingAssetMetadata;
 
   const feeValueView = new ValueView({
     valueView: {
       case: 'knownAssetId',
       value: {
         amount: fee?.amount ?? { hi: 0n, lo: 0n },
-        // TODO: once https://github.com/penumbra-zone/web/pull/1468 is merged, change this to metadata: assetFeeMedata
-        metadata: stakingAssetMetadata,
+        metadata: feeMetadata,
       },
     },
   });