+++ apps/minifront/src/components/swap/swap-form/token-swap-input.tsx
@@ -4,9 +4,13 @@ import { Metadata } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/
 import { Box } from '@repo/ui/components/ui/box';
 import { CandlestickPlot } from '@repo/ui/components/ui/candlestick-plot';
 import { joinLoHiAmount } from '@penumbra-zone/types/amount';
-import { getAmount, getBalanceView } from '@penumbra-zone/getters/balances-response';
+import {
+  getAmount,
+  getBalanceView,
+  getMetadataFromBalancesResponseOptional,
+} from '@penumbra-zone/getters/balances-response';
 import { ArrowRight } from 'lucide-react';
-import { useEffect } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { getBlockDate } from '../../../fetchers/block-date';
 import { AllSlices } from '../../../state';
 import { useStoreShallow } from '../../../utils/use-store-shallow';
@@ -26,6 +30,7 @@ import {
   swappableAssetsSelector,
   swappableBalancesResponsesSelector,
 } from '../../../state/swap/helpers';
+import { getDisplayDenomExponent } from '@penumbra-zone/getters/metadata';
 
 const getAssetOutBalance = (
   balancesResponses: BalancesResponse[] = [],
@@ -70,24 +75,49 @@ export const TokenSwapInput = () => {
   const { amount, setAmount, assetIn, setAssetIn, assetOut, setAssetOut, priceHistory, reverse } =
     useStoreShallow(tokenSwapInputSelector);
   const assetOutBalance = getAssetOutBalance(balancesResponses?.data, assetIn, assetOut);
+  const assetInExponent = useMemo(() => {
+    return getDisplayDenomExponent.optional()(getMetadataFromBalancesResponseOptional(assetIn));
+  }, [assetIn]);
 
-  useEffect(() => {
-    if (!assetIn || !assetOut) {
-      return;
-    } else {
-      return priceHistory.load();
-    }
-  }, [assetIn, assetOut]);
+  /**
+   * @todo this memo, state, and effects below are awkward. but this prevents
+   * excessive state churn. this is a temporary solution and will be corrected
+   * after implementing use of zquery for this data.
+   *
+   * @issue https://github.com/penumbra-zone/web/issues/1530
+   */
+  const loadPriceHistoryMemo = useMemo(
+    () => (assetIn && assetOut ? priceHistory.load : undefined),
+    [assetIn, assetOut, priceHistory.load],
+  );
 
+  const [updatedHeight, setUpdatedHeight] = useState<bigint>();
+
+  // initial price history
   useEffect(() => {
-    if (!priceHistory.candles.length) {
+    if (updatedHeight ?? !loadPriceHistoryMemo) {
       return;
-    } else if (latestKnownBlockHeight % 10n) {
-      return;
-    } else {
-      return priceHistory.load();
     }
-  }, [priceHistory, latestKnownBlockHeight]);
+    setUpdatedHeight(latestKnownBlockHeight);
+    return loadPriceHistoryMemo();
+  }, [assetIn, assetOut, latestKnownBlockHeight, loadPriceHistoryMemo, updatedHeight]);
+
+  // update price history
+  useEffect(() => {
+    // don't attempt to update an absent price history
+    if (!updatedHeight || !loadPriceHistoryMemo) {
+      return;
+    }
+
+    // rate limit pricing updates
+    const distanceFromLatest = latestKnownBlockHeight - updatedHeight;
+    if (!distanceFromLatest || distanceFromLatest < 10n) {
+      return;
+    }
+
+    setUpdatedHeight(latestKnownBlockHeight);
+    return loadPriceHistoryMemo();
+  }, [priceHistory.candles, latestKnownBlockHeight, loadPriceHistoryMemo, updatedHeight]);
 
   const maxAmount = getAmount.optional()(assetIn);
   const maxAmountAsString = maxAmount ? joinLoHiAmount(maxAmount).toString() : undefined;
@@ -108,6 +138,7 @@ export const TokenSwapInput = () => {
           variant='transparent'
           placeholder='Enter an amount...'
           max={maxAmountAsString}
+          maxExponent={assetInExponent}
           step='any'
           className={'font-bold leading-10 md:h-8 md:text-xl xl:h-10 xl:text-3xl'}
           onChange={e => {