+++ apps/minifront/src/state/ibc-out.ts
@@ -14,7 +14,7 @@ import {
 } from '@penumbra-zone/getters/value-view';
 import { getAddressIndex } from '@penumbra-zone/getters/address-view';
 import { toBaseUnit } from '@penumbra-zone/types/lo-hi';
-import { amountMoreThanBalance, planBuildBroadcast } from './helpers';
+import { amountMoreThanBalance, isIncorrectDecimal, planBuildBroadcast } from './helpers';
 import { getAssetId } from '@penumbra-zone/getters/metadata';
 import { assetPatterns } from '@penumbra-zone/types/assets';
 import { bech32, bech32m } from 'bech32';
@@ -233,6 +233,9 @@ export const ibcValidationErrors = (state: AllSlices) => {
     amountErr: !state.ibcOut.selection
       ? false
       : amountMoreThanBalance(state.ibcOut.selection, state.ibcOut.amount),
+    exponentErr: !state.ibcOut.selection
+      ? false
+      : isIncorrectDecimal(state.ibcOut.selection, state.ibcOut.amount),
   };
 };