+++ packages/protobuf/src/penumbra-core.ts
@@ -1,11 +1,13 @@
 export { QueryService as AppService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/app/v1/app_connect.js';
 
 export { QueryService as AuctionService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/component/auction/v1/auction_connect.js';
+export { QueryService as CommunityPoolService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/component/community_pool/v1/community_pool_connect.js';
 export { QueryService as CompactBlockService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/component/compact_block/v1/compact_block_connect.js';
 export {
   QueryService as DexService,
   SimulationService,
 } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/component/dex/v1/dex_connect.js';
+export { QueryService as FeeService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/component/fee/v1/fee_connect.js';
 export { QueryService as GovernanceService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/component/governance/v1/governance_connect.js';
 export { QueryService as SctService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/component/sct/v1/sct_connect.js';
 export { QueryService as ShieldedPoolService } from '@buf/penumbra-zone_penumbra.connectrpc_es/penumbra/core/component/shielded_pool/v1/shielded_pool_connect.js';