+++ packages/services/src/view-service/fees.ts
@@ -39,5 +39,30 @@ export const extractAltFee = async (
     return swaps?.swap?.claimFee?.assetId ?? indexedDb.stakingTokenAssetId;
   }
 
+  const auctionScheduleAsset = request.dutchAuctionScheduleActions
+    .map(a => a.description?.input)
+    .find(Boolean);
+  if (auctionScheduleAsset?.assetId) {
+    return auctionScheduleAsset.assetId;
+  }
+
+  const auctionEndAsset = request.dutchAuctionEndActions.map(a => a.auctionId).find(Boolean);
+  if (auctionEndAsset) {
+    const endAuction = await indexedDb.getAuction(auctionEndAsset);
+    if (endAuction.auction?.input?.assetId) {
+      return endAuction.auction.input.assetId;
+    }
+  }
+
+  const auctionWithdrawAsset = request.dutchAuctionWithdrawActions
+    .map(a => a.auctionId)
+    .find(Boolean);
+  if (auctionWithdrawAsset) {
+    const withdrawAuction = await indexedDb.getAuction(auctionWithdrawAsset);
+    if (withdrawAuction.auction?.input?.assetId) {
+      return withdrawAuction.auction.input.assetId;
+    }
+  }
+
   throw new Error('Could not extract alternative fee assetId from TransactionPlannerRequest');
 };