+++ packages/client/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @penumbra-zone/client
 
+## 12.0.0
+
+### Patch Changes
+
+- Updated dependencies [22bf02c]
+  - @penumbra-zone/protobuf@5.5.0
+
+## 11.1.1
+
+### Patch Changes
+
+- ab09596: fix manifest checking function
+
+## 11.1.0
+
+### Minor Changes
+
+- 72d83ea: provide type guards for event, and better types for event interface
+
 ## 11.0.1
 
 ### Patch Changes