+++ packages/crypto/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @penumbra-zone/crypto-web
 
+## 14.0.0
+
+### Patch Changes
+
+- @penumbra-zone/types@16.0.0
+
+## 13.0.1
+
+### Patch Changes
+
+- 3aaead1: Move the "default" option in package.json exports field to the last
+- Updated dependencies [3aaead1]
+  - @penumbra-zone/types@15.1.1
+
 ## 13.0.0
 
 ### Patch Changes