+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @penumbra-zone/ui
 
+## 7.0.2
+
+### Patch Changes
+
+- @penumbra-zone/types@16.0.0
+
+## 7.0.1
+
+### Patch Changes
+
+- Updated dependencies [3aaead1]
+  - @penumbra-zone/types@15.1.1
+
 ## 7.0.0
 
 ### Major Changes