+++ packages/wasm/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @penumbra-zone/wasm
 
+## 18.0.0
+
+### Patch Changes
+
+- Updated dependencies [22bf02c]
+  - @penumbra-zone/protobuf@5.5.0
+  - @penumbra-zone/types@16.0.0
+
+## 17.0.2
+
+### Patch Changes
+
+- Updated dependencies [3aaead1]
+  - @penumbra-zone/types@15.1.1
+
+## 17.0.1
+
+### Patch Changes
+
+- 1a57749: Bug fix with get_all_notes not respecting None asset id + delegator voting tests
+
 ## 17.0.0
 
 ### Minor Changes